rpi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyzmq>=25.1.0
psutil>=5.9.0
<<<<<<< HEAD
pyserial>=3.5
=======
flatbuffers>=23.5.26
pyserial>=3.5
aiohttp>=3.9.0
>>>>>>> 5376269 (rebase)
ELM327-emulator>=1.1.1
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
gpiozero>=1.6.2; platform_machine == "armv7l" or platform_machine == "aarch64"
